- Fixed warehouse map (removed temporary solution)
- Fixed malloca_s on release mode
- Added safety for SafeMemory on debug mode
- Fixed memory leak in EuclidRoom (loading)
- Added Mesh visualizer

+++ code/src/Graphics/Portals/Portal.cpp
@@ -13,7 +13,7 @@ Plutonium::Portal::Portal(Mesh * mesh)
 {
 	/* Calculate portal center. */
 	for (size_t i = 0; i < mesh->GetVertexCount(); i++) center += mesh->GetVertexAt(i).Position;
-	center /= mesh->GetVertexCount();
+	center /= static_cast<float>(mesh->GetVertexCount());
 
 	mesh->Finalize();
 }
